+++ bots/HodgePodge/modules/BotModule.py
@@ -1,9 +1,9 @@
-class Module():
+class BotModule():
     def __init__(self, name):
         self.name = name
     def getTriggerList(self):
         return []
     def connectDb(self, db):
         return
-    def respond(self, client, channel):
+    async def respond(self, client, channel):
         return