+++ bots/HodgePodge/docs/changelog.md
@@ -1,4 +1,5 @@
 # V.2
 
 Complete overhal of V.1 which was built in like 1 night and hacked with more features every other week. Cool new Features:
-
+1. Dice Rolls now support modifiers
+2.